adds navigation to detail view
