+++ TravelApp/Localizable.xcstrings
@@ -24,6 +24,9 @@
     "Animate" : {
 
     },
+    "asdasd" : {
+
+    },
     "Banner Image" : {
 
     },
